Auto-detect fits and show information about them
This mega-commit adds all our fits in a .dat file, then adds code to
detect each one of them (by identifying the tank on the fits). We then
detect the differences (eg. upgrades, missing modules, etc), and decide
whether or not we can approve it. Information about this process is also
shown in the UI.

The "ELITE-FIT" tag is given to anyone flying the elite fit, which
combined with elite/gold skills turns into the appropriate badge on the
UI. Implant badges are also shown.