Rework skills page to be more color-based
The old page had four tiers of colors, to explain the four upgrade tiers
on a five-level system in EVE. This made for a very colorful and useless
color scheme.

In the new way we simply show the four upgrade tiers as the four colors
from the scheme, with an indicator to show the next upgrade level. Much
cleaner, since we don't need to show the tier selector anymore!

(Checked with a color blind simulator and the colors don't seem to be so
close to each other that you can't read the page anymore.)