Stop blocking auto-approval based on not having minimum skills
To fly with TDF you need to online everything on the fit, and to have
the right tank skills. With both now checked separately from the basic
skills check, we can drop that check's impact on auto-approval.

Logi does have a much more extensive list of skill requirements and
will continue to require manual approval in case of missing skills.